+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add links to paths in admin app
 - remove use of fstrings to maintain v3.5 compatibility
 - minor tweaks for custom collections
+- return status 500 for application exceptions
 
 ### Added
 ## [6.14.0] - 2020-12-21